LayoutTests/platform/mac/accessibility/aria-popup.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div id="button" role="button" aria-haspopup="true" tabindex="0" style="width:200px; border: outset 2px gray; padding:3px">Example</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that when an ARIA button has aria-haspopup=true, it will be a popup button");

    if (window.accessibilityController) {

          document.getElementById("button").focus();
          var button = accessibilityController.focusedElement;

          shouldBe("button.role", "'AXRole: AXPopUpButton'");
          shouldBe("button.isPressActionSupported()", "true");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
